requirements.txt
Django>=4.0,<5.0
line-bot-sdk>=1.18.0,<2.0
openai>=0.10.0,<1.0
python-dotenv>=0.15.0,<1.0
mysqlclient>=2.0.3,<3.0